CM-62 Create form for creating interventions
