+++ package.json
@@ -50,15 +50,20 @@
     "@types/morgan": "^1.9.0",
     "@types/nunjucks": "^3.1.3",
     "@types/react": "^16.9.34",
+    "@types/react-datepicker": "^2.11.0",
     "@types/react-dom": "^16.9.6",
     "@types/react-plotly.js": "^2.2.4",
     "@types/webpack": "^4.41.10",
     "@types/webpack-dev-middleware": "^3.7.0",
     "bootstrap": "^4.4.1",
     "compression": "^1.7.4",
+    "css-loader": "^3.5.3",
     "express": "^4.17.1",
+    "formik": "^2.1.4",
     "morgan": "^1.10.0",
     "nunjucks": "^3.2.1",
+    "react-datepicker": "^2.16.0",
+    "style-loader": "^1.2.1",
     "ts-node": "^8.8.1",
     "typescript": "^3.8.3",
     "webpack": "^4.42.1",
