+++ webpack.config.ts
@@ -1,5 +1,5 @@
-import { Configuration } from "webpack";
-import * as path from "path";
+import * as path from 'path';
+import { Configuration } from 'webpack';
 
 let config: Configuration = {
   entry: "./frontend/index.ts",
@@ -11,19 +11,23 @@ let config: Configuration = {
         use: {
           loader: "ts-loader",
           options: {
-            configFile: "frontend/tsconfig.json"
-          }
+            configFile: "frontend/tsconfig.json",
+          },
         },
-        exclude: /node_modules/
-      }
-    ]
+        exclude: /node_modules/,
+      },
+      {
+        test: /\.css$/i,
+        use: ["style-loader", "css-loader"],
+      },
+    ],
   },
   resolve: {
-    extensions: [".ts", ".tsx", ".js"]
+    extensions: [".ts", ".tsx", ".js"],
   },
   output: {
     filename: "bundle.js",
-    path: path.resolve(__dirname, "dist/")
+    path: path.resolve(__dirname, "dist/"),
   },
   externals: {
     jQuery: "jQuery",
@@ -34,8 +38,8 @@ let config: Configuration = {
     "chroma-js": "chroma",
     react: "React",
     "react-dom": "ReactDOM",
-    "react-plotly.js": "createPlotlyComponent.default(Plotly)"
-  }
+    "react-plotly.js": "createPlotlyComponent.default(Plotly)",
+  },
 };
 
 export default config;
